requirements-freeze.txt
absl-py==2.3.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
asyncssh==2.21.1
atpublic==7.0.0
attrs==25.4.0
billiard==4.2.4
cachetools==6.2.2
ccxt==4.5.14
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.10
dvc==3.60.1
dvc-data==3.16.12
dvc-gs==3.0.2
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
filelock==3.20.0
flatbuffers==25.9.23
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gast==0.7.0
gcsfs==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grandalf==0.8
grpcio==1.76.0
gto==1.9.0
h5py==3.15.1
hydra-core==1.3.2
idna==3.11
iterative-telemetry==0.0.10
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.9
kombu==5.6.1
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.7.0
namex==0.1.0
narwhals==2.13.0
networkx==3.6
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scikit-learn==1.4.2
scipy==1.16.3
scmrepo==3.5.9
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==1.14.2
yarl==1.22.0
zc.lockfile==4.0
